# Conflicts:
#	src/main/kotlin/team5/backoffice/domain/course/service/CourseService.kt
#	src/main/kotlin/team5/backoffice/domain/lecture/service/LectureService.kt
#	src/main/kotlin/team5/backoffice/domain/review/service/ReviewService.kt

+++ src/main/kotlin/team5/backoffice/domain/course/service/CourseService.kt
@@ -12,6 +12,8 @@ import team5.backoffice.domain.course.repository.CategoryRepository
 import team5.backoffice.domain.course.repository.CourseRepository.CourseRepository
 import team5.backoffice.domain.course.repository.SubscriptionRepository
 import team5.backoffice.domain.review.repository.ReviewRepository
+import team5.backoffice.domain.exception.ModelNotFoundException
+import team5.backoffice.domain.exception.UnauthorizedUserException
 import team5.backoffice.domain.user.repository.StudentRepository
 import team5.backoffice.domain.user.repository.TutorRepository
 import java.time.LocalDateTime
@@ -38,7 +40,8 @@ class CourseService(
     }
 
     fun getCourseById(courseId: Long, studentId: Long?): CourseResponse {
-        val course = courseRepository.findByIdOrNull(courseId) ?: throw RuntimeException("Course not found")
+        val course =
+            courseRepository.findByIdOrNull(courseId) ?: throw ModelNotFoundException("course", "id: $courseId")
         course.increaseViewCount()
         return if (studentId != null) {
             CourseResponse.from(course, isBookmarkExists(courseId, studentId), isSubscribeExists(courseId, studentId), getCourseAverageRate(courseId))
@@ -56,15 +59,21 @@ class CourseService(
     fun createCourse(request: CourseRequest): CourseSimpleResponse {
 //        val category = categoryRepository.findByName(request.category) ?: throw RuntimeException("Category not found")
         val tutor = tutorRepository.findByIdOrNull(request.tutorId) ?: throw RuntimeException("Tid not found")
+    @Transactional
+    fun createCourse(request: CourseRequest, tutorId: Long): CourseSimpleResponse {
+        val category = categoryRepository.findByName(request.category) ?: throw ModelNotFoundException(
+            "category",
+            "name: ${request.category}"
+        )
+        val tutor = tutorRepository.findByIdOrNull(tutorId) ?: throw ModelNotFoundException("tutor", "id: $tutorId")
         return courseRepository.save(
             Course(
                 title = request.title,
                 tutor = tutor,
                 description = request.description,
-                category = request.category,
+                category = category,
                 imageUrl = request.imageUrl,
                 viewCount = 0,
-                createdAt = LocalDateTime.now()
             )
         ).let { CourseSimpleResponse.from(it) }
     }
@@ -74,6 +83,14 @@ class CourseService(
         val course = courseRepository.findByIdOrNull(courseId) ?: throw RuntimeException("Course not found")
 //        val category = categoryRepository.findByName(request.category) ?: throw RuntimeException("Category not found")
         if (course.tutor.id != request.tutorId) throw RuntimeException("Unauthorized tutor")
+    fun updateCourseById(courseId: Long, request: CourseRequest, tutorId: Long): CourseSimpleResponse {
+        val course =
+            courseRepository.findByIdOrNull(courseId) ?: throw ModelNotFoundException("course", "id: $courseId")
+        val category = categoryRepository.findByName(request.category) ?: throw ModelNotFoundException(
+            "category",
+            "name: ${request.category}"
+        )
+        if (course.tutor.id != tutorId) throw UnauthorizedUserException()
         course.apply {
             this.title = request.title
             this.description = request.description
@@ -85,28 +102,36 @@ class CourseService(
 
     fun deleteCourseById(courseId: Long) {
         val course = courseRepository.findByIdOrNull(courseId) ?: throw RuntimeException("Course not found")
+    @Transactional
+    fun deleteCourseById(courseId: Long, tutorId: Long) {
+        val course =
+            courseRepository.findByIdOrNull(courseId) ?: throw ModelNotFoundException("course", "id: $courseId")
+        if (course.tutor.id != tutorId) throw UnauthorizedUserException()
         courseRepository.delete(course)
     }
 
+    @Transactional
     fun addBookmark(courseId: Long, studentId: Long) {
-        courseRepository.findByIdOrNull(courseId) ?: throw RuntimeException("Course not found")
-        studentRepository.findByIdOrNull(studentId) ?: throw RuntimeException("Student not found")
+        courseRepository.findByIdOrNull(courseId) ?: throw ModelNotFoundException("course", "id: $courseId")
+        studentRepository.findByIdOrNull(studentId) ?: throw ModelNotFoundException("student", "id: $studentId")
         if (!isBookmarkExists(courseId, studentId)) {
             bookmarkRepository.save(Bookmark(BookmarkId(courseId, studentId)))
         }
     }
 
+    @Transactional
     fun removeBookmark(courseId: Long, studentId: Long) {
-        courseRepository.findByIdOrNull(courseId) ?: throw RuntimeException("Course not found")
-        studentRepository.findByIdOrNull(studentId) ?: throw RuntimeException("Student not found")
+        courseRepository.findByIdOrNull(courseId) ?: throw ModelNotFoundException("course", "id: $courseId")
+        studentRepository.findByIdOrNull(studentId) ?: throw ModelNotFoundException("student", "id: $studentId")
         val bookmark = bookmarkRepository.findByIdOrNull(BookmarkId(courseId, studentId))
-            ?: throw RuntimeException("Bookmark not found")
+            ?: throw ModelNotFoundException("bookmark", "id")
         bookmarkRepository.delete(bookmark)
     }
 
+    @Transactional
     fun subscribe(courseId: Long, studentId: Long) {
-        courseRepository.findByIdOrNull(courseId) ?: throw RuntimeException("Course not found")
-        studentRepository.findByIdOrNull(studentId) ?: throw RuntimeException("Student not found")
+        courseRepository.findByIdOrNull(courseId) ?: throw ModelNotFoundException("course", "id: $courseId")
+        studentRepository.findByIdOrNull(studentId) ?: throw ModelNotFoundException("student", "id: $studentId")
         if (!isSubscribeExists(courseId, studentId)) {
             subscriptionRepository.save(Subscription(SubscriptionId(courseId, studentId)))
         }
